basil-search/requirements.txt
# Web scraping and content processing
requests==2.31.0
beautifulsoup4==4.12.2
markdownify==0.11.6
lxml==4.9.3

# AI and NLP
sentence-transformers==2.2.2

# Database
chromadb==0.4.15

# Web API
fastapi==0.104.1
uvicorn==0.24.0

# Environment and configuration
python-dotenv==1.0.0

# Utilities
pathlib2==2.3.7

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.10.1
flake8==6.1.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Legacy dependencies (keeping for compatibility)
flask
fpdf2
markdown
spacy